+++ components/Cards/CardContracts.js
@@ -1,0 +1,86 @@
+import React from "react";
+
+// components
+
+export default function CardContracts() {
+  return (
+    <>
+      <div className="relative flex flex-col min-w-0 break-words bg-white w-full mb-6 shadow-lg rounded">
+        <div className="rounded-t mb-0 px-4 py-3 border-0">
+          <div className="flex flex-wrap items-center">
+            <div className="relative w-full px-4 max-w-full flex-grow flex-1">
+              <h3 className="font-semibold text-base text-gray-800">
+                Contracts
+              </h3>
+            </div>
+            <div className="relative w-full px-4 max-w-full flex-grow flex-1 text-right">
+              <button
+                className="bg-indigo-500 text-white active:bg-indigo-600 text-xs font-bold uppercase px-3 py-1 rounded outline-none focus:outline-none mr-1 mb-1 ease-linear transition-all duration-150"
+                type="button"
+              >
+                Deploy
+              </button>
+            </div>
+          </div>
+        </div>
+        <div className="block w-full overflow-x-auto" style={{ height: "296px"}}>
+          {/* Projects table */}
+          <table className="items-center w-full bg-transparent border-collapse">
+            <thead>
+              <tr>
+                <th className="px-6 bg-gray-100 text-gray-600 align-middle border border-solid border-gray-200 py-3 text-xs uppercase border-l-0 border-r-0 whitespace-no-wrap font-semibold text-left">
+                  Contract name
+                </th>
+                <th className="px-6 bg-gray-100 text-gray-600 align-middle border border-solid border-gray-200 py-3 text-xs uppercase border-l-0 border-r-0 whitespace-no-wrap font-semibold text-left">
+                  Address
+                </th>
+                <th className="px-6 bg-gray-100 text-gray-600 align-middle border border-solid border-gray-200 py-3 text-xs uppercase border-l-0 border-r-0 whitespace-no-wrap font-semibold text-left">
+                  Created At
+                </th>
+                <th className="px-6 bg-gray-100 text-gray-600 align-middle border border-solid border-gray-200 py-3 text-xs uppercase border-l-0 border-r-0 whitespace-no-wrap font-semibold text-left">
+                  Last Block
+                </th>
+              </tr>
+            </thead>
+            <tbody>
+              <tr>
+                <th className="border-t-0 px-6 align-middle border-l-0 border-r-0 text-sm whitespace-no-wrap p-4 text-left text-blue-500">
+                  Shield.sol
+                </th>
+                <td className="border-t-0 px-6 align-middle border-l-0 border-r-0 text-xs whitespace-no-wrap p-4">
+                  <a href="https://goerli.etherscan.io/address/0x63758bc241d4cd924ebfbed273a2f6a1179f8f86" target="_blank" className="text-black hover:text-green-500 font-semibold">
+                      0x63758bc241d4cd924ebfbed273a2f6a1179f8f86
+                      </a>
+                </td>
+                <td className="border-t-0 px-6 align-middle border-l-0 border-r-0 text-xs whitespace-no-wrap p-4">
+                  2021-01-29
+                </td>
+                <td className="border-t-0 px-6 align-middle border-l-0 border-r-0 text-xs whitespace-no-wrap p-4">
+                  <i className="fa fa-cubes text-green-500 mr-4"></i>
+                  4192259
+                </td>
+              </tr>
+              <tr>
+                <th className="border-t-0 px-6 align-middle border-l-0 border-r-0 text-sm whitespace-no-wrap p-4 text-left text-blue-500">
+                  VerifierNoop.sol
+                </th>
+                <td className="border-t-0 px-6 align-middle border-l-0 border-r-0 text-xs whitespace-no-wrap p-4">
+                <a href="https://goerli.etherscan.io/address/0x76f272ba2b1c3887f117dfeeb600e53b50a2207b" target="_blank" className="text-black hover:text-green-500 font-semibold">
+                    0x76f272ba2b1c3887f117dfeeb600e53b50a2207b
+                    </a>
+                </td>
+                <td className="border-t-0 px-6 align-middle border-l-0 border-r-0 text-xs whitespace-no-wrap p-4">
+                  2021-01-29
+                </td>
+                <td className="border-t-0 px-6 align-middle border-l-0 border-r-0 text-xs whitespace-no-wrap p-4">
+                  <i className="fa fa-cubes text-green-500 mr-4"></i>
+                  4192263
+                </td>
+              </tr>
+            </tbody>
+          </table>
+        </div>
+      </div>
+    </>
+  );
+}