+++ components/Navbars/AuthNavbar.js
@@ -38,10 +38,10 @@ export default function Navbar(props) {
               <li className="flex items-center">
                 <a
                   className="lg:text-white lg:hover:text-gray-300 text-gray-800 px-3 py-4 lg:py-2 flex items-center text-xs uppercase font-bold"
-                  href="https://www.creative-tim.com/learning-lab/tailwind/nextjs/overview/notus?ref=nnjs-auth-navbar"
+                  href="https://github.com/sekmet/Baseline-Chain-Dashboard/blob/main/CHANGELOG.md"
                 >
                   <i className="lg:text-gray-300 text-gray-500 far fa-file-alt text-lg leading-lg mr-2" />{" "}
-                  Docs
+                  Changelog
                 </a>
               </li>
             </ul>