+++ pages/_app.js
@@ -54,26 +54,6 @@ const getEthChainId = async () => {
 
 export default class MyApp extends App {
   componentDidMount = async () => {
-    let comment = document.createComment(`
-
-      =========================================================
-      * Notus NextJS - v1.0.0 based on Tailwind Starter Kit by Creative Tim
-      =========================================================
-
-      * Product Page: https://www.creative-tim.com/product/notus-nextjs
-      * Copyright 2020 Creative Tim (https://www.creative-tim.com)
-      * Licensed under MIT (https://github.com/creativetimofficial/notus-nextjs/blob/master/LICENSE.md)
-
-      * Tailwind Starter Kit Page: https://www.creative-tim.com/learning-lab/tailwind-starter-kit/presentation
-
-      * Coded by Creative Tim
-
-      =========================================================
-
-      * The above copyright notice and this permission notice shall be included in all copies or substantial portions of the Software.
-    `);
-    document.insertBefore(comment, document.documentElement);
-
     this.setState({
       ...this.state,
       chainId: await getEthChainId(),