+++ pages/admin/didgenerator.js
@@ -1,0 +1,24 @@
+import React from "react";
+
+// components
+
+import CardSettings from "components/Cards/CardSettings.js";
+import CardProfile from "components/Cards/CardProfile.js";
+
+// layout for page
+
+import Admin from "layouts/Admin.js";
+
+export default function Didgenerator() {
+  return (
+    <>
+      <div className="flex flex-wrap">
+        <div className="w-full lg:w-12/12 px-4">
+          <CardSettings />
+        </div>
+      </div>
+    </>
+  );
+}
+
+Didgenerator.layout = Admin;