+++ pages/admin/settings.js
@@ -2,19 +2,29 @@ import React from "react";
 
 // components
 
-import CardSettings from "components/Cards/CardSettings.js";
-import CardProfile from "components/Cards/CardProfile.js";
+import FormSettings from "components/Forms/FormSettings.js";
 
 // layout for page
 
 import Admin from "layouts/Admin.js";
 
-export default function Settings() {
+import commitMgrConfig from '../../configs/commit_mgr.config';
+
+// This function gets called at build time
+export async function getStaticProps() {
+  // By returning { props: { posts } }, the Blog component
+  // will receive `posts` as a prop at build time
+  return {
+    props: { ...commitMgrConfig },
+  }
+}
+
+export default function Settings(props) {
   return (
     <>
       <div className="flex flex-wrap">
         <div className="w-full lg:w-12/12 px-4">
-          <CardSettings />
+          <FormSettings {...props} />
         </div>
       </div>
     </>