+++ pages/auth/login.js
@@ -22,16 +22,16 @@ export default function Login() {
       <div className="container mx-auto px-4 h-full">
         <div className="flex content-center items-center justify-center h-full">
           <div className="w-full lg:w-6/12 px-4">
-            <div className="relative flex flex-col min-w-0 break-words w-full mb-6 shadow-lg rounded-lg bg-gray-300 border-0">
+            <div className="relative flex flex-col min-w-0 break-words w-full mb-6 shadow-lg rounded-lg bg-indigo-500 border-0">
               <div className="rounded-t mb-0 px-6 py-6">
                 <div className="text-center mb-3">
                   <img src="https://www.baseline-protocol.org/images/baseline_Horizontal-Logo-allWhite.svg" />
-                  <h3 className="text-gray-600 text-sm font-bold">
+                  <h3 className="text-white text-sm font-bold">
                     Connect with
                   </h3>
                 </div>
                 {wallet.status === 'connected' ? (
-                  <div className="text-sm text-gray-600 font-semibold py-1 text-center md:text-left">
+                  <div className="text-sm text-white font-semibold py-1 text-center md:text-left">
                     <h6>Account: {wallet.account}</h6>
                     <h6>Balance: {wallet.balance}</h6>
                     <button
@@ -42,7 +42,7 @@ export default function Login() {
                       <img
                         alt="..."
                         className="w-5 mr-1"
-                        src={require("assets/img/github.svg")}
+                        src={require("assets/img/logoff.svg")}
                       />
                       Disconnect
                     </button>
@@ -58,7 +58,7 @@ export default function Login() {
                       <img
                         alt="..."
                         className="w-5 mr-1"
-                        src={require("assets/img/github.svg")}
+                        src={require("assets/img/key.svg")}
                       />
                       MetaMask
                     </button>
@@ -70,7 +70,7 @@ export default function Login() {
                     <img
                       alt="..."
                       className="w-5 mr-1"
-                      src={require("assets/img/google.svg")}
+                      src={require("assets/img/shield.svg")}
                     />
                     Frame
                   </button>
@@ -82,14 +82,14 @@ export default function Login() {
                     <img
                       alt="..."
                       className="w-5 mr-1"
-                      src={require("assets/img/google.svg")}
+                      src={require("assets/img/key.svg")}
                     />
                     Portis
                   </button>
                   </div>
                 )}
                 <hr className="mt-6 border-b-1 border-gray-400" />
-                <h6 className="text-center mt-3"><strong>Network: </strong>{wallet.networkName}</h6>
+                <h6 className="text-center text-white mt-3"><strong>Network: </strong>{wallet.networkName}</h6>
               </div>
             </div>
           </div>